backend/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
gunicorn
psycopg2-binary
requests

beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
